Refactoring: move SimpleSpectrumData to model, remove dia_common, change intensities precision to f32 in rt/intensity pairs
* SimpleSpectrumData refactoring:
- Add SimpleSpectrumData to model.rs with DataPointProvider impl and From<SpectrumData> conversions
- Add get_simple_spectrum_data() in queries.rs using proper encoding-aware BB parsing
- Update diafication.rs: use MzDbReader + get_simple_spectrum_data() instead of hardcoded parsing

* dia_common removal:
- Remove unused calculate_mz_bounds_from_arrays function from dia_common.rs
- Inline dia_common.rs utilities into diafication.rs (DiaWriteContext, DiaSpectrumParams, etc.)
- Delete writer/dia_common.rs (was only used by diafication, not dia_simplifier)

* Change RtIntensityPair from (f32, f64) to (f32, f32):
- Update SignalSmoother trait and impls for f32 intensity
- Update PeakelFinder trait and impls (Basic, Smart, Histogram) for f32 intensity
- Update BaselineRemover threshold methods for f32
- Update Peakel.elution_time_intensity_pairs() to return Vec<(f32, f32)>

+++ mzdb-rs/src/processing/model.rs
@@ -272,7 +272,7 @@ impl LcContext {
 // ============================================================================
 
 /// A simple time-intensity pair for XIC data
-pub type RtIntensityPair = (f32, f64);
+pub type RtIntensityPair = (f32, f32);
 
 /// Collection of RT-intensity pairs
 pub type RtIntensityPairs = Vec<RtIntensityPair>;
@@ -426,11 +426,11 @@ impl Peakel {
     }
 
     /// Get RT-intensity pairs
-    pub fn elution_time_intensity_pairs(&self) -> Vec<(f32, f64)> {
+    pub fn elution_time_intensity_pairs(&self) -> Vec<(f32, f32)> {
         self.elution_times
             .iter()
             .zip(self.intensity_values.iter())
-            .map(|(&rt, &intensity)| (rt, intensity as f64))
+            .map(|(&rt, &intensity)| (rt, intensity))
             .collect()
     }
 }
